Lead scoring case study for X- education
This case study has several objectives:

Create a logistic regression model to award each lead a score between 0 and 100. A higher score indicates a hot lead that is most likely to convert, while a lower value indicates a cold lead that is unlikely to convert. The firm may bring other issues that your model should be prepared to address if the company's needs alter in the future.